// app/pro-plus/model/order-history.ts
/**
 * Beacon New Rest Service Ver 1.0
 * Release notes:  `4.8`  * Add `onHold` to request body of endpoint `/submitOrder`
 *
 * OpenAPI spec version: release/4.8
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Job } from './job';
import { OrderHistoryShipping } from './order-history-shipping';

export interface OrderHistory {
    total?: number;
    sellingBranch?: string;
    accountId?: string;
    orderNumber?: string;
    job?: Job;
    purchaseOrderNo?: string;
    creditHold?: string;
    shipping?: OrderHistoryShipping;
    orderPlacedDate: string;
    otherCharges?: number;
    tax?: number;
    /**
     * Order Status
     */
    orderStatusCode?: OrderHistoryObj.OrderStatusCodeEnum;
    subTotal?: number;
    orderId: string;
    specialInstruction: string;
    accountToken?: string;
}
export namespace OrderHistoryObj {
    export type OrderStatusCodeEnum = 'I' | 'R' | 'P' | 'C' | 'O' | 'K' | 'N';
    export const OrderStatusCodeEnum = {
        I: 'I' as OrderStatusCodeEnum,
        R: 'R' as OrderStatusCodeEnum,
        P: 'P' as OrderStatusCodeEnum,
        C: 'C' as OrderStatusCodeEnum,
        O: 'O' as OrderStatusCodeEnum,
        K: 'K' as OrderStatusCodeEnum,
        N: 'N' as OrderStatusCodeEnum,
    };
}

export interface OrderHistoryV2 {
    total?: number;
    sellingBranch?: string;
    accountId?: string;
    orderNumber?: string;
    job?: {
        jobName: string;
        jobNumber: string;
    };
    purchaseOrderNo?: string;
    creditHold?: string;
    shipping?: {
        shippingBranch: number;
        shippingBranchDisplayName: string;
        address: {
            postalCode: string;
            state: string;
            address1: string;
            address2: string;
            address3: string;
            country: string;
            city: string;
        };
        shippingMethod: string;
    };
    orderPlacedDate?: string;
    tax?: number;
    orderStatusCode?: OrderHistoryObj.OrderStatusCodeEnum;
    subTotal?: number;
    orderId: string;
    specialInstruction: string;

    //  New
    totalNull: boolean;
    displayTotalMsg: boolean;
    totalMsg: string;
    inStore: string;
    source: string;
    taxNull: boolean;
    orderStatusValue: string;
    subTotalNull: boolean;
    showDeliveryNotification: boolean;
    showShipTips: boolean;
    shipTipsMsg: string;
    otherCharges?: number;
}